Standardize Socco formatting for Scio examples
