+++ scio-examples/src/main/scala/com/spotify/scio/examples/cookbook/BigQueryTornadoes.scala
@@ -21,7 +21,7 @@
 // `sbt runMain "com.spotify.scio.examples.cookbook.BigQueryTornadoes
 // --project=[PROJECT] --runner=DataflowRunner --zone=[ZONE]
 // --input=clouddataflow-readonly:samples.weather_stations
-// --output=[DATASET].bigquery_tornadoes`
+// --output=[DATASET].bigquery_tornadoes"`
 package com.spotify.scio.examples.cookbook
 
 import com.google.api.services.bigquery.model.{TableFieldSchema, TableSchema}
