+++ scio-examples/src/main/scala/com/spotify/scio/examples/extra/AvroInOut.scala
@@ -16,6 +16,11 @@
  */
 
 // Example: Avro Input and Output
+// Usage:
+
+// `sbt runMain "com.spotify.scio.examples.extra.AvroInOut
+// --project=[PROJECT] --runner=DataflowRunner --zone=[ZONE]
+// --input=[INPUT].avro --output=[OUTPUT].avro --method=[METHOD]"`
 package com.spotify.scio.examples.extra
 
 import com.spotify.scio._
