+++ scio-examples/src/main/scala/com/spotify/scio/examples/extra/CheckpointExample.scala
@@ -15,6 +15,12 @@
  * under the License.
  */
 
+// Example: Use Checkpoints to debug a long workflow
+// Usage:
+
+// `sbt runMain "com.spotify.scio.examples.extra.CheckpointExample
+// --project=[PROJECT] --runner=DataflowRunner --zone=[ZONE]
+// --checkpoint=gs://[CHECKPOINT_FILE] --output=[OUTPUT]"`
 package com.spotify.scio.examples.extra
 
 import com.spotify.scio.ContextAndArgs
