+++ scio-examples/src/main/scala/com/spotify/scio/examples/extra/MetricsExample.scala
@@ -19,6 +19,10 @@
 // scalastyle:off regex
 
 // Example: Metrics Example
+// Usage:
+
+// `sbt runMain "com.spotify.scio.examples.extra.MetricsExample
+// --project=[PROJECT] --runner=DataflowRunner --zone=[ZONE]"`
 package com.spotify.scio.examples.extra
 
 import com.spotify.scio._
