+++ scio-examples/src/main/scala/com/spotify/scio/examples/extra/ShapelessAvroExample.scala
@@ -15,6 +15,11 @@
  * under the License.
  */
 
+// Example: Shapeless Avro examples
+// Avro examples using shapeless-datatype to seamlessly convert between case classes and Avro
+// generic records.
+
+// See: https://github.com/nevillelyh/shapeless-datatype
 package com.spotify.scio.examples.extra
 
 import com.spotify.scio._
@@ -23,12 +28,6 @@ import com.spotify.scio.avro._
 import com.spotify.scio.examples.common.ExampleData
 import org.apache.avro.generic.GenericRecord
 
-/*
- * Avro examples using shapeless-datatype to seamlessly convert between case classes and Avro
- * generic records.
- *
- * https://github.com/nevillelyh/shapeless-datatype
- */
 object ShapelessAvroExample {
   // limit import scope to avoid polluting namespace
   import shapeless.datatype.avro._
@@ -38,16 +37,13 @@ object ShapelessAvroExample {
   case class WordCount(word: String, count: Long)
 }
 
-/*
-SBT
-runMain
-  com.spotify.scio.examples.extra.ShapelessAvroWriteExample
-  --project=[PROJECT] --runner=DataflowRunner --zone=[ZONE]
-  --input=gs://apache-beam-samples/shakespeare/kinglear.txt
-  --output=gs://[BUCKET]/[PATH]/wordcount-avro
- */
+// ## Count words and save result to Avro
 
-// Count words and save result to Avro
+// Usage:
+// `sbt runMain "com.spotify.scio.examples.extra.ShapelessAvroWriteExample
+// --project=[PROJECT] --runner=DataflowRunner --zone=[ZONE]
+// --input=gs://apache-beam-samples/shakespeare/kinglear.txt
+// --output=gs://[BUCKET]/[PATH]/wordcount-avro"`
 object ShapelessAvroWriteExample {
   def main(cmdlineArgs: Array[String]): Unit = {
     import ShapelessAvroExample._
@@ -63,16 +59,13 @@ object ShapelessAvroWriteExample {
   }
 }
 
-/*
-SBT
-runMain
-  com.spotify.scio.examples.extra.ShapelessAvroReadExample
-  --project=[PROJECT] --runner=DataflowRunner --zone=[ZONE]
-  --input=gs://[BUCKET]/[PATH]/wordcount-avro
-  --output=gs://[BUCKET]/[PATH]/wordcount
- */
+// ## Read word count result back from Avro
+// Usage:
 
-// Read word count result back from Avro
+// `sbt runMain "com.spotify.scio.examples.extra.ShapelessAvroReadExample
+// --project=[PROJECT] --runner=DataflowRunner --zone=[ZONE]
+// --input=gs://[BUCKET]/[PATH]/wordcount-avro
+// --output=gs://[BUCKET]/[PATH]/wordcount"`
 object ShapelessAvroReadExample {
   def main(cmdlineArgs: Array[String]): Unit = {
     import ShapelessAvroExample._
