+++ scio-examples/src/main/scala/com/spotify/scio/examples/extra/SingleGZipFileExample.scala
@@ -15,6 +15,12 @@
  * under the License.
  */
 
+// Example: Gzip File write
+// Usage:
+
+// `sbt runMain "com.spotify.scio.examples.extra.SingleGZipFileExample
+// --project=[PROJECT] --runner=DataflowRunner --zone=[ZONE]
+// --input=[INPUT.txt] --output=[OUTPUT]"`
 package com.spotify.scio.examples.extra
 
 import com.spotify.scio.ContextAndArgs
