+++ scio-examples/src/main/scala/com/spotify/scio/examples/extra/TapExamples.scala
@@ -15,14 +15,22 @@
  * under the License.
  */
 
-// scalastyle:off regex
-
-// Example: Handling Output with Tap and Future
+// Example: Handling I/O with Tap and Future
 package com.spotify.scio.examples.extra
 
 import com.spotify.scio._
+import com.spotify.scio.examples.common.ExampleData
+import com.spotify.scio.io.Taps
 
-object TapExample {
+// ## Tap Output example
+// Save collections as text files, then open their `Tap`s in a new `ScioContext`
+
+// Usage:
+
+// `sbt runMain "com.spotify.scio.examples.extra.TapOutExample
+// --project=[PROJECT] --runner=DataflowRunner --zone=[ZONE]
+// --output=gs://[OUTPUT] --method=[METHOD]"`
+object TapOutputExample {
   def main(cmdlineArgs: Array[String]): Unit = {
     // Each `ScioContext` instance maps to a unique pipeline
 
@@ -41,10 +49,11 @@ object TapExample {
       .saveAsTextFile(args("output"))
     sc1.close()
 
-    // Wait for future completions, which should happen when `sc1` finishes
+    // Wait for `Future` completions, which should happen when `sc1` finishes
     val t1 = f1.waitForResult()
     val t2 = f2.waitForResult()
 
+    // scalastyle:off regex
     // Fetch `Tap` values directly
     println(t1.value.mkString(", "))
     println(t2.value.mkString(", "))
@@ -57,5 +66,40 @@ object TapExample {
     val result = sc2.close().waitUntilFinish()
 
     println(result.finalState)
+    // scalastyle:on regex
+  }
+}
+
+// ## Tap Input example
+// Use `Tap`s as input to defer execution logic until both `Future`s complete
+
+// Usage:
+
+// `sbt runMain "com.spotify.scio.examples.extra.TapInputExample
+// --project=[PROJECT] --runner=DataflowRunner --zone=[ZONE]"`
+object TapInputExample {
+  def main(cmdlineArgs: Array[String]): Unit = {
+    import scala.concurrent.ExecutionContext.Implicits.global
+    val taps = Taps() // entry point to acquire taps
+
+    // extract `Tap[T]`s from two `Future[Tap[T]]`s
+    val r = for {
+      t1 <- taps.textFile(ExampleData.KING_LEAR)
+      t2 <- taps.textFile(ExampleData.OTHELLO)
+    } yield {
+      // execution logic when both taps are available
+      val (sc, _) = ContextAndArgs(cmdlineArgs)
+      val out = (t1.open(sc) ++ t2.open(sc))
+        .flatMap(_.split("[^a-zA-Z']+").filter(_.nonEmpty))
+        .countByValue
+        .map(kv => kv._1 + "\t" + kv._2)
+        .materialize
+      sc.close()
+      out
+    }
+
+    // scalastyle:off regex
+    println(r.waitForResult().value.take(10).toList)
+    // scalastyle:on regex
   }
 }
