+++ scio-examples/src/main/scala/com/spotify/scio/examples/extra/TsvExample.scala
@@ -15,13 +15,14 @@
  * under the License.
  */
 
+// Example: Reading and writing tsv data
 package com.spotify.scio.examples.extra
 
 import com.spotify.scio._
 import com.spotify.scio.examples.common.ExampleData
 import org.apache.beam.sdk.{io => beam}
 
-// Example: Word Count Example writing to TSV file
+// ## Writing TSV data example
 // Usage:
 
 // `sbt runMain "com.spotify.scio.examples.extra.TsvExampleWrite
@@ -37,7 +38,7 @@ object TsvExampleWrite {
     val input = args.getOrElse("input", ExampleData.KING_LEAR)
     val output = args("output")
 
-    // Create a Write Transformation with tsv suffix, single file and specific header.
+    // Create a `Write` Transformation with tsv suffix, single file and specific header.
     val transform = beam.TextIO
       .write()
       .to(pathWithShards(output))
@@ -65,7 +66,7 @@ object TsvExampleWrite {
     path.replaceAll("\\/+$", "") + "/part"
 }
 
-// # Reading TSV data Example
+// ## Reading TSV data example
 // Usage:
 
 // `sbt runMain "com.spotify.scio.examples.extra.TsvExampleRead
