+++ scio-examples/src/main/scala/com/spotify/scio/examples/extra/TypedBigQueryTornadoes.scala
@@ -20,7 +20,7 @@
 
 // `sbt runMain "com.spotify.scio.examples.extra.TypedBigQueryTornadoes
 // --project=[PROJECT] --runner=DataflowRunner --zone=[ZONE]
-// --output=[PROJECT:DATASET.TABLE]`
+// --output=[PROJECT]:[DATASET].[TABLE]"`
 package com.spotify.scio.examples.extra
 
 import com.spotify.scio.bigquery._
@@ -29,22 +29,22 @@ import com.spotify.scio.ContextAndArgs
 object TypedBigQueryTornadoes {
 
   // Annotate input class with schema inferred from a BigQuery SELECT.
-  // Class Row will be expanded into a case class with fields from the SELECT query. A companion
+  // Class `Row` will be expanded into a case class with fields from the SELECT query. A companion
   // object will also be generated to provide easy access to original query/table from annotation,
-  // TableSchema and converter methods between the generated case class and TableRow.
+  // `TableSchema` and converter methods between the generated case class and `TableRow`.
   @BigQueryType.fromQuery("SELECT tornado, month FROM [publicdata:samples.gsod]")
   class Row
 
   // Annotate output case class.
   // Note that the case class is already defined and will not be expanded. Only the companion
-  // object will be generated to provide easy access to TableSchema and converter methods.
+  // object will be generated to provide easy access to `TableSchema` and converter methods.
   @BigQueryType.toTable
   case class Result(month: Long, tornado_count: Long)
 
   def main(cmdlineArgs: Array[String]): Unit = {
     val (sc, args) = ContextAndArgs(cmdlineArgs)
 
-    // Get input from BigQuery and convert elements from TableRow to Row.
+    // Get input from BigQuery and convert elements from `TableRow` to `Row`.
     // SELECT query from the original annotation is used by default.
     sc.typedBigQuery[Row]()
       .flatMap(r => if (r.tornado.getOrElse(false)) Seq(r.month) else Nil)
